separate types to halve header files
