--- src/EnableFlag.hpp
@@ -1,9 +1,0 @@
-#pragma once
-
-enum MGLEnableFlag {
-	MGL_NOTHING = 0,
-	MGL_BLEND = 1,
-	MGL_DEPTH_TEST = 2,
-	MGL_CULL_FACE = 4,
-	MGL_ENABLE_MASK = 7,
-};
