+++ src/Error.cpp
@@ -1,4 +1,4 @@
-#include "Error.hpp"
+#include "Types.hpp"
 
 #include <cstdarg>
