--- src/InvalidObject.hpp
@@ -1,9 +1,0 @@
-#pragma once
-
-#include "Python.hpp"
-
-struct MGLInvalidObject {
-	PyObject_HEAD
-};
-
-extern PyTypeObject MGLInvalidObject_Type;
