+++ src/ModernGL.cpp
@@ -1,29 +1,4 @@
-#include "Attribute.hpp"
-#include "Buffer.hpp"
-#include "BufferAccess.hpp"
-#include "ComputeShader.hpp"
-#include "Context.hpp"
-#include "EnableFlag.hpp"
-#include "Error.hpp"
-#include "Framebuffer.hpp"
-#include "InvalidObject.hpp"
-#include "Primitive.hpp"
-#include "Program.hpp"
-#include "ProgramStage.hpp"
-#include "Renderbuffer.hpp"
-#include "Shader.hpp"
-#include "Subroutine.hpp"
-#include "SubroutineUniform.hpp"
-#include "Texture.hpp"
-#include "TextureCube.hpp"
-#include "Texture3D.hpp"
-#include "TextureFilter.hpp"
-#include "Uniform.hpp"
-#include "UniformBlock.hpp"
-#include "Varying.hpp"
-#include "VertexArray.hpp"
-#include "VertexArrayAttribute.hpp"
-#include "VertexArrayListAttribute.hpp"
+#include "Types.hpp"
 
 #include "GLContext.hpp"
