+++ src/Texture3D.cpp
@@ -1,8 +1,4 @@
-#include "Texture3D.hpp"
-
-#include "Error.hpp"
-#include "InvalidObject.hpp"
-#include "Buffer.hpp"
+#include "Types.hpp"
 
 #include "InlineMethods.hpp"
