--- src/VertexArrayListAttribute.hpp
@@ -1,12 +1,0 @@
-#pragma once
-
-#include "Python.hpp"
-
-struct MGLVertexArrayListAttribute {
-	PyObject_HEAD
-
-	PyObject * content;
-	int location;
-};
-
-extern PyTypeObject MGLVertexArrayListAttribute_Type;
